Popups actually close now when the page changes.
Before adding the ability to use the browser's back button, the user had to click on the back button provided with the UI, which caused any popup panes to close. The same functionality has been added to the browser back button, so that any time a different page is opened, any previous popups are closed.